spatial-engine/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
shapely>=2.0.0
geopandas>=0.13.0
pandas>=2.0.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
celery>=5.3.0
redis>=4.5.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
